refactor: make `database` package public
