+++ internal/reporter/report_test.go
@@ -2,7 +2,7 @@ package reporter_test
 
 import (
 	"osv-detector/internal"
-	"osv-detector/internal/database"
+	"osv-detector/pkg/database"
 	"osv-detector/pkg/lockfile"
 	"osv-detector/internal/reporter"
 	"strings"
