+++ pkg/database/osv_test.go
@@ -2,7 +2,7 @@ package database_test
 
 import (
 	"osv-detector/internal"
-	"osv-detector/internal/database"
+	"osv-detector/pkg/database"
 	"osv-detector/pkg/lockfile"
 	"testing"
 	"time"
